+++ src/main/java/com/barbershop/api/Models/Relations/ShopBarber.java
@@ -1,15 +1,16 @@
 package com.barbershop.api.Models.Relations;
 
 import com.barbershop.api.Models.BaseModel;
+
 import javax.persistence.Entity;
 import java.util.Date;
 
 @Entity
 public class ShopBarber extends BaseModel {
-    public Integer shopId, barberId;
+    public Long shopId, barberId;
     public String photo, location, schedule;
 
-    public ShopBarber(long id, boolean active, Integer shopId, Integer barberId, String photo, String location, String schedule) {
+    public ShopBarber(long id, boolean active, Long shopId, Long barberId, String photo, String location, String schedule) {
         super(id, active);
         this.shopId = shopId;
         this.barberId = barberId;
@@ -21,7 +22,7 @@ public class ShopBarber extends BaseModel {
     public ShopBarber() {
     }
 
-    public ShopBarber(boolean active, Integer shopId, Integer barberId, String photo, String location, String schedule) {
+    public ShopBarber(boolean active, Long shopId, Long barberId, String photo, String location, String schedule) {
         super(active);
         this.shopId = shopId;
         this.barberId = barberId;
@@ -30,7 +31,7 @@ public class ShopBarber extends BaseModel {
         this.schedule = schedule;
     }
 
-    public ShopBarber(long id, boolean active, Date createdUtc, Date modifiedUtc, Integer shopId, Integer barberId, String photo, String location, String schedule) {
+    public ShopBarber(long id, boolean active, Date createdUtc, Date modifiedUtc, Long shopId, Long barberId, String photo, String location, String schedule) {
         super(id, active, createdUtc, modifiedUtc);
         this.shopId = shopId;
         this.barberId = barberId;
@@ -39,19 +40,19 @@ public class ShopBarber extends BaseModel {
         this.schedule = schedule;
     }
 
-    public Integer getShopId() {
+    public Long getShopId() {
         return shopId;
     }
 
-    public void setShopId(int shopId) {
+    public void setShopId(Long shopId) {
         this.shopId = shopId;
     }
 
-    public Integer getBarberId() {
+    public Long getBarberId() {
         return barberId;
     }
 
-    public void setBarberId(int barberId) {
+    public void setBarberId(Long barberId) {
         this.barberId = barberId;
     }
